Perfect! Now, lets add another round. This will actually be the first round. We'll need to explain how each round works at the beginning of the round.
This new round will work as follow. The user will see a word written in a different color, the same way our current test works. However, in this round, the correct answer is actually the color corresponding to the word, rather than the color of the font. That is, if the word "Purple" is written in green, the correct answer is purple